% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFormulaTable.R
\name{buildFormulaTable}
\alias{buildFormulaTable}
\title{Reads and builds FormulaTable file into a data.frame}
\usage{
buildFormulaTable(x = NULL, atom.symbol = "C")
}
\arguments{
\item{x}{File directory (as in \emph{utils::read.table})}

\item{atom.symbol}{Atom symbol of the labelling experiment, for C13 set it to \code{"C"}}
}
\value{
formulaTable with NumAtoms column
}
\description{
Reads text file and creates \code{NumAtoms} column by obtaining carbon number from compound formula
}

