% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaTables-data.R
\docType{data}
\name{formulaTables}
\alias{formulaTables}
\title{Example data frames input for autoQ function}
\format{A data.frame containing minimal information to run \link{autoQ} with extdata files}
\usage{
data(formulaTables)
}
\description{
Example data frames input for autoQ function
}
\examples{
data(formulaTables)
}
\keyword{datasets}
