% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanRawPlot.R
\name{meanRawPlot}
\alias{meanRawPlot}
\title{Read all scans raw data, calculate average spectrum for all files and plot}
\usage{
meanRawPlot(SampleFiles = NULL, formulaTable = NULL, RTwin = 5, topdf = NULL)
}
\arguments{
\item{SampleFiles}{List of files to be plotted (see \emph{base::list.files()})}

\item{formulaTable}{Data.frame including CompoundName, Formula, mz, RT and NumAtoms (see \link{buildFormulaTable})}

\item{RTwin}{Retention time window (in seconds) to extract the metabolite isotopologues}

\item{topdf}{File directory where the plots will be saved in pdf format. One pdf file per mzML file in Sampledir}
}
\value{
Plots per each file and compound including all isotopologues
}
\description{
Raw data plotting average spectra of all SampleFiles and compound for quality control pourposes
}
\note{
To plot files separately, use \link{rawPlot}
}
