% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metBarPlot.R
\name{metBarPlot}
\alias{metBarPlot}
\title{Creates a PDF file with barplots per each compound}
\usage{
metBarPlot(x = NULL, topdf = NULL, val.to.plot = "Area", groups = c(),
  ylabel = NULL, ...)
}
\arguments{
\item{x}{The data.frame result from \link{autoQ} or \link{QTransform}}

\item{topdf}{File directory where the plot should be saved. If \code{NULL} plots will be sent to R plotting device}

\item{val.to.plot}{Values to use for plotting. Either "Area" or "Maxo". If values were transformed with \code{Qtransform}, indicate the value used for transformation}

\item{groups}{Sample groups. A vector of length equal to number of samples that will match the samples order}

\item{ylabel}{Label for y axis in the plot. If not indicated, val.to.plot value will be used as default.}

\item{...}{other arguments that will be passed to \emph{grDevices::pdf}}
}
\value{
Barplots per each compound including deviation in each isotopologue.
}
\description{
Creates a PDF file with barplots (Mean +- SD) per each compound and isotopologues using ggplot2
}

