% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTransform.R
\name{QTransform}
\alias{QTransform}
\title{Area or Maxo value transformation}
\usage{
QTransform(x = NULL, val.to.use = "Area", val.trans = c("P", "R"))
}
\arguments{
\item{x}{The data.frame result from \link{autoQ}}

\item{val.to.use}{Type of quantification to transform. Either "Area" or "Maxo"}

\item{val.trans}{Set it to "P" for labelling percentages or "R" for relative quantification to the Monoisotopic peak(for isotopic pattern observation)}
}
\value{
A data.frame with transformed values. Can be input to \link{metBarPlot}.
}
\description{
Creates a data.frame per each compound and isotopologues calculating required transformation, prior to export or plotting
}

