% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawPlot.R
\name{rawPlot}
\alias{rawPlot}
\title{Read all scans raw data and plot}
\usage{
rawPlot(SampleFiles = NULL, formulaTable = NULL, RTwin = 5,
  topdf = NULL, plotall = F)
}
\arguments{
\item{SampleFiles}{List of files to be plotted (see \emph{base::list.files()})}

\item{formulaTable}{Data.frame including CompoundName, Formula, mz, RT and NumAtoms (see \link{buildFormulaTable})}

\item{RTwin}{Retention time window (in seconds) to extract the metabolite isotopologues}

\item{topdf}{File directory where the plots will be saved in pdf format. One pdf file per mzML file in Sampledir}
}
\value{
Plots per each file and compound including all isotopologues
}
\description{
Raw data plotting per each file and compound for quality control pourposes
}
\note{
It is recommended to run it for a few files only, plotting a high number of files may take time.
}

