% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplifybyPpm}
\alias{simplifybyPpm}
\title{Simplify data.frame of quantified isotopologues}
\usage{
simplifybyPpm(finalres, ppm = NULL)
}
\arguments{
\item{finalres}{Dataframe resulting from \emp{autoQ} function from high-resolution data.}

\item{ppm}{Relative mass tolerance to cluster isotopologue masses}
}
\value{
A data.frame with a reduced number of rows in which similar mass isotopologues have been collapsed into a single row.
}
\description{
Simplify data.frame of quantified isotopologues by clustering isotopologue ions within a certain ppm mass difference. 
These differences are caused by variation in recorded ion m/z through that may not match from sample to sample.
}
