% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subNatural.R
\name{subNatural}
\alias{subNatural}
\title{Corrects natural abundance}
\usage{
subNatural(autoQres)
}
\arguments{
\item{autoQres}{The data.frame result from \link{autoQ}}
}
\value{
A data frame with a row per each isotopologue and compound with corrected abundances per sample.
}
\description{
Corrects natural abundance given the abundance predicted within autoQ. This abundance is calculated by enviPat functions.
}
