% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subNatural.R
\name{subNaturalM0}
\alias{subNaturalM0}
\title{Corrects natural abundance}
\usage{
subNaturalM0(autoQres)
}
\arguments{
\item{autoQres}{The data.frame result from \link{autoQ}}
}
\value{
A data frame with a row per each isotopologue and compound with corrected abundances per sample.
}
\description{
Corrects natural abundance given the abundance predicted within autoQ, using only M+0 isotopologue.
This abundance is previously calculated by enviPat functions.
}
