% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumIsotopologues.R
\name{sumIsotopologues}
\alias{sumIsotopologues}
\title{Sums equivalent isotopologues abundances}
\usage{
sumIsotopologues(autoQres)
}
\arguments{
\item{autoQres}{The data.frame result from \link{autoQ}}
}
\value{
A data frame with a row per each isotopologue and compound with summed abundances per sample.
}
\description{
Sums abundances of ions equivalent to each isotopologue (M+x). Necessary to execute metBarPlot function 
after running autoQ function on high-resolution data.
}
