% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoQ.R
\name{autoQ}
\alias{autoQ}
\title{Read files and quantify each isotopologue}
\usage{
autoQ(SampleFiles = NULL, formulaTable = NULL, IsoMass = 1.003355,
  RTwin = 5, toplot = F, ScanE = 0.1, IntE = 0.4, fit.p = 0.05)
}
\arguments{
\item{SampleFiles}{List of files to be processed (see \emph{base::list.files()})}

\item{formulaTable}{Data.frame including CompoundName, Formula, mz, RT and NumAtoms (see \link{buildFormulaTable})}

\item{IsoMass}{Stable isotope mass difference, for C13: 1.003355}

\item{RTwin}{Retention time window (in seconds) to extract the metabolite isotopologues}

\item{ScanE}{Absolute error in which to search the Maxo scan}

\item{IntE}{Absolute error in which to integrate the area of the isotopologue. Should be larger than \emph{ScanE} and smaller than 0.5}

\item{fit.p}{Max linear model p-value for a peak to have its area quantified, otherwise only Maxo is returned}
}
\value{
A data.frame including the Area and Maxo for each file and metabolite isotopologues. If peak was missing, values are NA.
}
\description{
Read each file and quantify both Area and Maxo of each metabolite isotopologues, by evaluating the RT window of each metabolite to minimize background noise
}

